// frontend/src/api/generated/types.gen.ts
export type RefreshResponse = {
    /**
     * New JWT access token
     */
    accessToken?: string;
    /**
     * New refresh token (old token is now invalid)
     */
    refreshToken?: string;
    /**
     * Token type
     */
    tokenType?: string;
    /**
     * Access token expiration time in seconds
     */
    expiresInSeconds?: number;
};
// ...
export type RefreshResponses = {
    /**
     * Token refreshed successfully
     */
    200: RefreshResponse;
};

export type RefreshResponse2 = RefreshResponses[keyof RefreshResponses];